+++ tweaks/sources/sudo-tweak/remove.sh
@@ -14,12 +14,19 @@ if [ ! -e /usr/local/etc/sudoers.d/nodns ] || [ ! -e /usr/local/etc/sudoers.d/no
 	echo 'Not installed'
 	exit 1
 fi
+if [ ! -e /usr/local/etc/sudoers.d/nolecture ] || [ ! -e /usr/local/etc/sudoers.d/nolecture ]; then
+	echo 'Not installed'
+	exit 1
+fi
 
 echo -n '[rm] /usr/local/etc/sudoers.d/nodns'
 	rm /usr/local/etc/sudoers.d/nodns > /dev/null 2>&1 && echo ' [OK]' || echo ' [Fail]'
+echo -n '[rm] /usr/local/etc/sudoers.d/nolecture'
+	rm /usr/local/etc/sudoers.d/nolecture > /dev/null 2>&1 && echo ' [OK]' || echo ' [Fail]'
 
-cd /etc/sudoers.d
 echo -n '[rm] /etc/sudoers.d/nodns'
 	rm /etc/sudoers.d/nodns > /dev/null 2>&1 && echo ' [OK]' || echo ' [Fail]']'
+echo -n '[rm] /etc/sudoers.d/nolecture'
+	rm /etc/sudoers.d/nolecture > /dev/null 2>&1 && echo ' [OK]' || echo ' [Fail]']'
 
 exit 0
